+++ lib/dart_store.dart
@@ -2,15 +2,12 @@ library dart_store;
 
 import 'package:meta/meta.dart';
 import 'package:eventify/eventify.dart' show Event, EventEmitter;
-
+import 'package:mixed_collection/mixed_collection.dart';
 part 'src/core/model.dart';
 part 'src/core/store.dart';
 part 'src/factory/storefactory.dart';
 part 'src/factory/modelfactory.dart';
 part 'src/config.dart';
-part 'src/core/database_error.dart';
 part 'src/core/storage.dart';
-part 'src/core/filterable.dart';
-part 'src/core/sortable.dart';
 part 'src/stores/memorystore.dart';
-part 'src/core/collection.dart';
+part 'src/core/modelcollection.dart';